fix: 결제 가능한 테이블 없을시 예외 처리 추가
