+++ src/main/java/Application.java
@@ -1,27 +1,14 @@
-import controller.PosController;
-import domain.*;
-import view.InputView;
+import controller.PosMachine;
 import view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
-        final Tables tables = new Tables(TableRepository.tables());
-        final Menus menus = new Menus(MenuRepository.menus());
-        Pos pos;
-
-        do {
-            pos = getPosWithValidation();
-            PosController controller = pos.getController();
-            controller.controlAction(tables, menus);
-        } while (pos.isNotExit());
-    }
-
-    private static Pos getPosWithValidation() {
         try {
-            return Pos.getPos(InputView.inputPosNumber());
-        } catch (IllegalArgumentException e) {
+            PosMachine posMachine = new PosMachine();
+            posMachine.run();
+        } catch (IllegalArgumentException | NullPointerException e) {
             OutputView.printExceptionMessage(e.getMessage());
-            return getPosWithValidation();
+            main(args);
         }
     }
 }
